fix: bounding box coords for 4.5.4-dev

+++ src/person.py
@@ -71,7 +71,7 @@ class PlayerTracker:
                 cv2.rectangle(frame, bbox, (0, 255, 0), 10)
             elif cv2.__version__ == '4.5.4-dev': 
                 bbox = track.to_tlbr()
-                cv2.rectangle(frame, bbox[0:2], bbox[2:], (0, 255, 0), 10)
+                cv2.rectangle(frame, bbox[0:2].astype(int), bbox[2:].astype(int), (0, 255, 0), 10)
             cv2.putText(frame, f'ID: {track.track_id}', (int(bbox[0]), int(bbox[1])-10), cv2.FONT_HERSHEY_SIMPLEX, 0.9, (0,255,0), 2)
         
         return frame, detections
